+++ tests/Component/HttpFoundation/QueryParameterBagTest.php
@@ -8,8 +8,10 @@
 
 namespace HeimrichHannot\RequestBundle\Test\Component\HttpFoundation;
 
+use Contao\CoreBundle\Routing\ScopeMatcher;
 use Contao\TestCase\ContaoTestCase;
 use HeimrichHannot\RequestBundle\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestMatcher;
 use Symfony\Component\HttpFoundation\RequestStack;
 
 class QueryParameterBagTest extends ContaoTestCase
@@ -39,7 +41,12 @@ class QueryParameterBagTest extends ContaoTestCase
         $requestStack = new RequestStack();
         $requestStack->push(new \Symfony\Component\HttpFoundation\Request());
 
-        $this->request = new Request($this->mockContaoFramework(), $requestStack);
+        $backendMatcher = new RequestMatcher('/contao');
+        $frontendMatcher = new RequestMatcher('/index');
+
+        $scopeMatcher = new ScopeMatcher($backendMatcher, $frontendMatcher);
+
+        $this->request = new Request($this->mockContaoFramework(), $requestStack, $scopeMatcher);
     }
 
     public function testAddUnusedParameters()