+++ CHANGELOG.md
@@ -4,4 +4,4 @@ All notable changes to this project will be documented in this file.
 ## [1.0.1] - 2018-03-12
 
 ### Fixed
-- composer dependencies and replacement for `heimrichhannot/contao-request`
+- removed `heimrichhannot/contao-utils-bundle` circular dependency